requirements.txt
appdirs==1.4.4
argh==0.26.2
asgiref==3.5.0
atomicwrites==1.4.0
bkcharts==0.2
black==19.10b0
boto==2.49.0
brotlipy==0.7.0
dask==2021.10.0
Django==3.2.12
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
entrypoints==0.3
et-xmlfile==1.1.0
fonttools==4.25.0
gmpy2==2.0.8
inflection==0.5.1
llvmlite==0.37.0
locket==0.2.1
Markdown==3.3.6
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
munkres==1.1.4
mypy-extensions==0.4.3
nltk==3.6.5
pandas==1.3.4
password-strength==0.0.3.post2
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==8.4.0
pkginfo==1.7.1
ply==3.11
psycopg2==2.9.3
pycosat==0.6.3
pycurl==7.44.1
PyJWT==2.3.0
pyls-spyder==0.4.0
pyodbc===4.0.0-unsupported
pytest==6.2.4
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
PyYAML==6.0
scikit-image==0.18.3
simplegeneric==0.8.1
sip==4.19.13
sqlparse==0.4.2
unicodecsv==0.14.1
urllib3==1.26.7
webencodings==0.5.1
xlwings==0.24.9
xlwt==1.3.0
zict==2.0.0
django-cors-headers==3.12.0
